+++ service-base/src/main/java/org/eclipse/hono/service/registration/RegistrationMessageFilter.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2016 Bosch Software Innovations GmbH.
+ * Copyright (c) 2016, 2017 Bosch Software Innovations GmbH.
  *
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -14,7 +14,6 @@ package org.eclipse.hono.service.registration;
 import org.apache.qpid.proton.amqp.messaging.AmqpValue;
 import org.apache.qpid.proton.message.Message;
 import org.eclipse.hono.util.BaseMessageFilter;
-import org.eclipse.hono.util.MessageHelper;
 import org.eclipse.hono.util.RegistrationConstants;
 import org.eclipse.hono.util.ResourceIdentifier;
 import org.slf4j.Logger;
@@ -45,17 +44,14 @@ public final class RegistrationMessageFilter extends BaseMessageFilter {
          } else if (msg.getMessageId() == null && msg.getCorrelationId() == null) {
              LOG.trace("message has neither a message-id nor correlation-id");
              return false;
-         } else if (!hasValidAction(msg)) {
+         } else if (!RegistrationConstants.isValidAction(msg.getSubject())) {
              LOG.trace("message [{}] does not contain valid action property", msg.getMessageId());
              return false;
          } else if (msg.getReplyTo() == null) {
              LOG.trace("message [{}] contains no reply-to address", msg.getMessageId());
              return false;
          } else if (msg.getBody() != null) {
-             if (msg.getContentType() == null || !msg.getContentType().startsWith("application/json")) {
-                 LOG.trace("message [{}] content type is not JSON", msg.getMessageId());
-                 return false;
-             } else if (!(msg.getBody() instanceof AmqpValue)) {
+             if (!(msg.getBody() instanceof AmqpValue)) {
                  LOG.trace("message [{}] contains non-AmqpValue section payload", msg.getMessageId());
                  return false;
              } else {
@@ -65,13 +61,4 @@ public final class RegistrationMessageFilter extends BaseMessageFilter {
              return true;
          }
     }
-
-     private static boolean hasValidAction(final Message msg) {
-         String action = MessageHelper.getApplicationProperty(
-                 msg.getApplicationProperties(),
-                 RegistrationConstants.APP_PROPERTY_ACTION,
-                 String.class);
-         return RegistrationConstants.isValidAction(action);
-     }
-
 }